@@ -1,6 +1,6 @@
 {
   "name": "reglini",
-  "version": "1.16.1",
+  "version": "1.16.2",
   "private": true,
   "scripts": {
     "dev": "next dev",
@@ -11,55 +11,60 @@
   "dependencies": {
     "@headlessui/react": "^1.7.17",
     "@heroicons/react": "^2.0.13",
-    "@prisma/client": "^5.3.1",
-    "@tanstack/react-query": "^4.18.0",
-    "@trpc/client": "^10.4.0",
-    "@trpc/next": "^10.4.0",
-    "@trpc/react-query": "^10.4.0",
-    "@trpc/server": "^10.4.0",
+    "@prisma/client": "^5.4.0",
+    "@t3-oss/env-nextjs": "^0.6.1",
+    "@tanstack/react-query": "^4.35.7",
+    "@trpc/client": "^10.38.5",
+    "@trpc/next": "^10.38.5",
+    "@trpc/react-query": "^10.38.5",
+    "@trpc/server": "^10.38.5",
     "ae_sdk": "^0.4.13",
-    "axios": "^1.4.0",
-    "bcrypt": "^5.1.0",
-    "cloudinary": "^1.32.0",
-    "dotenv": "^16.0.3",
+    "axios": "^1.5.1",
+    "bcrypt": "^5.1.1",
+    "cloudinary": "^1.41.0",
     "framer-motion": "^8.5.2",
-    "html-react-parser": "^3.0.4",
-    "js-cookie": "^3.0.1",
+    "html-react-parser": "^4.2.2",
+    "js-cookie": "^3.0.5",
     "multer": "^1.4.5-lts.1",
     "multer-storage-cloudinary": "^4.0.0",
-    "next": "13.0.5",
-    "next-auth": "^4.17.0",
+    "next": "^13.4.19",
+    "next-auth": "^4.23.0",
     "next-connect": "^0.13.0",
-    "next-intl": "^2.10.2",
+    "next-intl": "^2.20.2",
     "next-pwa": "^5.6.0",
     "next-themes": "^0.2.1",
-    "nodemailer": "^6.8.0",
+    "nodemailer": "^6.9.5",
     "nprogress": "^0.2.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "slugify": "^1.6.5",
+    "slugify": "^1.6.6",
     "superjson": "^1.11.0",
-    "zod": "^3.19.1",
-    "zustand": "^4.1.4"
+    "zod": "^3.21.4",
+    "zustand": "^4.4.2"
   },
   "devDependencies": {
     "@babel/eslint-parser": "^7.22.15",
     "@tailwindcss/aspect-ratio": "^0.4.2",
     "@tailwindcss/forms": "^0.5.3",
     "@types/bcrypt": "^5.0.0",
-    "@types/js-cookie": "^3.0.2",
-    "@types/lodash": "^4.14.191",
-    "@types/multer": "^1.4.7",
-    "@types/node": "^18.11.9",
-    "@types/nodemailer": "^6.4.6",
-    "@types/nprogress": "^0.2.0",
+    "@types/js-cookie": "^3.0.4",
+    "@types/lodash": "^4.14.199",
+    "@types/multer": "^1.4.8",
+    "@types/next-pwa": "^5.6.6",
+    "@types/node": "^20.8.2",
+    "@types/nodemailer": "^6.4.11",
+    "@types/nprogress": "^0.2.1",
     "@types/react": "^18.0.25",
     "@types/react-dom": "^18.0.9",
+    "@typescript-eslint/eslint-plugin": "^6.7.4",
+    "@typescript-eslint/parser": "^6.7.4",
     "autoprefixer": "^10.4.13",
     "eslint": "^8.50.0",
     "eslint-config-next": "^13.5.4",
-    "postcss": "^8.4.19",
-    "prisma": "^5.3.1",
+    "postcss": "^8.4.31",
+    "prettier": "^3.0.3",
+    "prettier-plugin-tailwindcss": "^0.5.5",
+    "prisma": "^5.4.0",
     "tailwindcss": "^3.3.3",
     "typescript": "5.2"
   }
