@@ -3,7 +3,7 @@ import { Listbox, Transition } from "@headlessui/react";
 import { CheckIcon, ChevronUpDownIcon } from "@heroicons/react/24/solid";
 import { useTranslations } from "next-intl";
 
-import { Wilaya } from "@prisma/client";
+import type { Wilaya } from "@prisma/client";
 import {
   SHADOW,
   PADDING,
@@ -12,7 +12,7 @@ import {
   TEXT_INPUT,
 } from "~/config/design";
 import { Loading } from "~/components/shared";
-import { trpc } from "~/utils/trpc";
+import { api } from "~/utils/api";
 
 export function SelectWilaya({
   wilaya,
@@ -21,7 +21,7 @@ export function SelectWilaya({
   wilaya?: Wilaya;
   setWilaya: Dispatch<SetStateAction<Wilaya | undefined>>;
 }) {
-  const wilayasQuery = trpc.address.wilayas.useQuery();
+  const wilayasQuery = api.address.wilayas.useQuery();
   const t = useTranslations("AccountPage.details.address");
 
   return (
