@@ -5,10 +5,10 @@ import { useTranslations } from "next-intl";
 import { PADDING, ROUNDED, SHADOW } from "~/config/design";
 import ItemProperties from "~/components/account/ItemProperties";
 import { Loading, Title } from "~/components/shared";
-import { trpc } from "~/utils/trpc";
+import { api } from "~/utils/api";
 
 const ListOrders = () => {
-  const ordersQuery = trpc.order.all.useQuery();
+  const ordersQuery = api.order.all.useQuery();
   const t = useTranslations("AccountPage.orders");
 
   return (
