@@ -4,7 +4,7 @@ import { Tab } from "@headlessui/react";
 import { PADDING } from "~/config/design";
 import ConvertCurrency from "~/components/currency/ConvertCurrency";
 import HistoricalRates from "~/components/currency/HistoricalRates";
-import LiveRate from "./LiveRate";
+import LiveRate from "~/components/currency/LiveRate";
 
 const TABS = [
   {
