@@ -3,13 +3,13 @@ import { ReactNode, useCallback, useEffect } from "react";
 import Navbar from "~/components/layout/Navbar";
 import Footer from "~/components/layout/Footer";
 import { useFinance } from "~/utils/store";
-import { trpc } from "~/utils/trpc";
+import { api } from "~/utils/api";
 
 const Layout = ({ children }: { children: ReactNode }) => {
   const { set_currency, set_commission, commission, euro, usd } = useFinance();
 
-  const commissionMutation = trpc.commission.useMutation();
-  const currenciesMutation = trpc.currency.currenciesObject.useMutation();
+  const commissionMutation = api.commission.useMutation();
+  const currenciesMutation = api.currency.currenciesObject.useMutation();
 
   const fetchFinance = useCallback(async () => {
     if (!commission)
